refactor(ipc): file and class renames
