+++ GlazeWM.Application.IpcServer/GlazeWM.Application.IpcServer.csproj
@@ -1,5 +1,4 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
     <TargetFramework>net6.0-windows10.0.17763.0</TargetFramework>
     <DebugType>embedded</DebugType>
@@ -12,5 +11,4 @@
       Include="NetCoreServer"
       Version="6.7.0" />
   </ItemGroup>
-
 </Project>
